[Feat] Debut implémentation confirmation de la commande : traitement par lignes des quantites en fonction des stocks restants. TODO: Traiter si le stock est déjà à 0 entre temps puis ajmise à jour ou suppression des lignes dans le panier si stock = 0

+++ src/main/java/fr/ut1/m2ipm/dafumarket/dao/PanierDAO.java
@@ -22,7 +22,7 @@ public class PanierDAO {
 
     }
 
-    public Optional<Panier> getPanierById(int id) {
+    public Optional<Panier> getPanierById(Long id) {
         return this.panierRepository.findById((long) id);
     }
